Can vote or not

+++ voters.java
@@ -1,7 +1,7 @@
 public class voters {
     public static void main(String[]args)
     {
-        float age=20;
+        int age=20;
         if(age>=18)
         {
             System.out.println("Can vote");
